backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.54.0
pillow==10.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
boto3==1.33.6
python-multipart==0.0.6
PyPDF2==3.0.1
aiofiles==23.2.1
httpx==0.27.2
